requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
asyncpg==0.18.3
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
discord.py==1.0.0a1634+g7f4c57d
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
oauthlib==3.0.0
psutil==5.4.8
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.7.0
urllib3==1.24.1
websockets==6.0
yarl==1.3.0
